scripts: by convention variables are capitalized
Environment variable names used by the utilities in the Shell and
Utilities volume of POSIX.1-2017 consist solely of uppercase letters,
digits, and the underscore ( _ ) from the characters defined in Portable
Character Set and do not begin with a digit.

...

The name space of environment variable names containing lowercase
letters is reserved for applications. Applications can define any
environment variables with names from this name space without modifying
the behavior of the standard utilities # Please enter the commit message
for your changes. Lines starting